feat(增加模糊查询)

+++ model/user.go
@@ -4,6 +4,6 @@ type User struct {
 	Id       int    `json:"userId"`
 	Username string `json:"nikename"`
 	Password string
-	Money    int `json:"money"`
+	Money    float64 `json:"money"`
 	Avatar   string
 }
